Still need a mechanism to detect env.reset()
Very close to finish line!
If total number of products >= 200, the program run normally (through it gradually get slower). But if total < 200, an environment reset will occur in main.py, but our get_action still stick with the old environment - which's just disappeared, therefore it malfunctions.
So, your mission is (in commit msg). An idea is to check for the first-used-stock's remaining area, if it (suddenly) increase, then you know the env is reset and call Policy.reset().